cleaner version
still stuck with the runaway fnl threads problem but at least this version is cleaner.

Basically what happens is that I can switch from thread 2 to thread 3 every 2000ms fine and then back to thread  2 but after that it goes into a 'frenzy' mode which basically the scheduler runs not the actual treads.